requirements.txt
astroid==3.1.0
attrs==23.2.0
black==24.2.0
click==8.1.7
dill==0.3.8
isort==5.13.2
libcst==1.2.0
loguru==0.7.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pylint==3.1.0
PyYAML==6.0.1
tomli==2.0.1
tomlkit==0.12.4
typing-inspect==0.9.0
typing_extensions==4.10.0
